% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_area.R
\name{bb_area}
\alias{bb_area}
\title{Calculates catchment area}
\usage{
bb_area(station_basins, germanyshape, polygon_col = "objectid")
}
\arguments{
\item{station_basins}{a list, each element has a vector with primary basin id from germanyshape}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{polygon_col}{name of id column in shapefile}
}
\value{
A vector with catchment area in km2 for each list element (i.e. each catchment)
}
\description{
Calculates catchment area
}
\examples{
\dontrun{
will follow
}
}
