% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_catchments.R
\name{bb_catchments}
\alias{bb_catchments}
\title{Creates one catchment polygon from basic basins, calculates catchment area and saves them as geopackage}
\usage{
bb_catchments(
  station_basins,
  germanyshape,
  polygon_col = "objectid",
  save = FALSE,
  dsn = "",
  layer = "none",
  stations = NULL,
  station_id_col = "stationid"
)
}
\arguments{
\item{station_basins}{a list, each element has a vector with primary basin id from germanyshape}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{polygon_col}{name of basin id column in shapefile}

\item{save}{boolean, if TRUE catchments of stations will be saved as a geopackage under \code{dsn}}

\item{dsn}{of basin id column in shapefile}

\item{layer}{layer name in geopackage to be saved, e.g. Bundesland (if save TRUE)}

\item{stations}{\code{sf} point object of stations, columns will be retained}

\item{station_id_col}{name of basin id column in shapefile}
}
\value{
\code{sf} polygon object with catchment area in km2 in column area_bb, columns from \code{stations} will be retained
}
\description{
Creates one catchment polygon from basic basins, calculates catchment area and saves them as geopackage
}
\examples{
\dontrun{}
will follow
}
