% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_catchments_par.R
\name{bb_catchments_par}
\alias{bb_catchments_par}
\title{Creates a catchment polygon from list of basic basins running in parallel and calculates area and saves a geopackage}
\usage{
bb_catchments_par(
  station_basins,
  ncores,
  germanyshape,
  polygon_col = "objectid",
  stations = NULL,
  station_id_col = "stationid",
  dissolve = TRUE,
  save = FALSE,
  dsn = "",
  layer = NULL
)
}
\arguments{
\item{station_basins}{a list, each element has a vector with primary basin id from germanyshape}

\item{ncores}{integer, number of cores to be used during parallel processing}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{polygon_col}{name of basin id column in shapefile}

\item{stations}{\code{sf} point object of stations, columns will be retained as polygon attributes, order of stations must be identical with the 'station_basins'}

\item{station_id_col}{name of basin id column in shapefile}

\item{dissolve}{boolean, if TRUE subbasins are dissolved into one catchment polygon}

\item{save}{boolean, if TRUE catchments of stations will be saved as a geopackage under 'dsn'}

\item{dsn}{path and file name of geopackage (if 'save' TRUE)}

\item{layer}{layer name within geopackage to be saved, e.g. Bundesland (if 'save' TRUE)}
}
\value{
\code{sf} polygon object with column area_bb for catchment area in km2, columns from 'stations' will be retained
}
\description{
Creates a catchment polygon from list of basic basins running in parallel and calculates area and saves a geopackage
}
\examples{
\dontrun{}
will follow
}
