% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_ids.R
\name{bb_ids}
\alias{bb_ids}
\title{Function to extract ID (polygon_id) for one or multiple gauging stations.}
\usage{
bb_ids(stations, shape, debug = FALSE, polygon_col = "polygon_id")
}
\arguments{
\item{stations}{data.frame with gauging station with x, y}

\item{shape}{path to main shapefile}

\item{debug}{boolean; prints working process}

\item{polygon_col}{name of id column in shapefile}
}
\value{
ID of basin that includes the gauging station.
}
\description{
Function to extract ID (polygon_id) for one or multiple gauging stations.
}
\details{
Stations must be a data.frame of one or more gauging stations with x and y (longitude, latitude in WGS84) columns.
}
