% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_readgermanyshape.R
\name{bb_readgermanyshape}
\alias{bb_readgermanyshape}
\title{Read basin shape with \code{sf}}
\usage{
bb_readgermanyshape(file = NULL, name = "shape", quiet = FALSE)
}
\arguments{
\item{file}{Location of the shapefile or geopackage file}

\item{name}{Variable name that the shapefile should be assigned to.}

\item{quiet}{logical; suppress info from \code{sf} package on name, driver, size and spatial reference etc.}
}
\value{
Loads the basins shape and stores it in a variable given by \code{name}.
The \code{sf} package will generate some printouts or warnings during processing.
}
\description{
Read basin shape with \code{sf} package and stores as global variable defined by \code{name}.
}
\examples{
gpkg_file <- '~/basisbasins.gpkg'
bb_readgermanyshape(file = gpkg_file, name = 'basins_germany')

#wrapper for this:
basins_germany <- sf::st_read('~/basisbasins.gpkg')
}
