% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_setbasinshape.R
\name{bb_setbasinshape}
\alias{bb_setbasinshape}
\title{Sets path and filename of the basin shape.}
\usage{
bb_setbasinshape(basinshapefile = NULL)
}
\arguments{
\item{basinshapefile}{Path and filename of the basin shape file.}
}
\value{
Sets a Global Environment Variable in R named \code{BB_BASINSHAPE}.
}
\description{
Sets path and filename of the basin shape.
}
\details{
The file must be in a readable format for the \code{sf} package, e.g. \code{.shp} or \code{.gpkg}
}
\examples{
bb_setbasinshape('~/basisbasins.shp')
Sys.getenv("BB_BASINSHAPE") # returns variable
}
