% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_stationid.R
\name{bb_stationid}
\alias{bb_stationid}
\title{Function to extract basin ID (polygon_id) for one or multiple gauging stations.}
\usage{
bb_stationid(stations, germanyshape, debug = FALSE, polygon_col = "polygon_id")
}
\arguments{
\item{stations}{data.frame with gauging station with x and y coordinates.}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{debug}{boolean; TRUE prints working process}

\item{polygon_col}{name of id column in shapefile}
}
\value{
Basin ID that includes the gauging station. For each line in \code{stations} one ID will be returned.
}
\description{
Function to extract basin ID (polygon_id) for one or multiple gauging stations.
}
\details{
Stations must be a data.frame of one or more gauging stations with x and y (longitude, latitude in WGS84) columns.
}
\examples{
# will follow
}
