% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_delineate.R
\name{bb_delineate}
\alias{bb_delineate}
\title{Delineation of catchment(s)}
\usage{
bb_delineate(
  station_ids,
  germanyshape,
  polygon_col = "objectid",
  geb = "geb_kz_num",
  gew = "gew_kz_num",
  remove_artifical = TRUE
)
}
\arguments{
\item{station_ids}{vector with primary basin IDs for one or more gauging stations}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{polygon_col}{name of primary basin ID column in shapefile}

\item{geb}{name of domain column ('Gebietskennzahl'), default = 'geb_kz_num')}

\item{gew}{name of water body column ('Gewässerkennzahl', default = 'gew_kz_num')}

\item{remove_artifical}{logical, should artificial water bodies such as canals be removed from procedure?}
}
\value{
will follow
}
\description{
Delineation of catchment(s)
}
\examples{
\dontrun{
will follow
}
}
