% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_stationid.R
\name{bb_stationid}
\alias{bb_stationid}
\title{Extract primary basin ID for one or multiple gauging stations}
\usage{
bb_stationid(stations, germanyshape, debug = FALSE, polygon_col = "polygon_id")
}
\arguments{
\item{stations}{\code{data.frame} with gauging station(s) with x and y coordinates. See details.}

\item{germanyshape}{basin shapefile as a \code{sf} object; can be generated with \code{\link[=bb_readgermanyshape]{bb_readgermanyshape()}}.}

\item{debug}{boolean; TRUE prints working process}

\item{polygon_col}{name of primary basin ID column in shapefile, e.g. \code{objectid}}
}
\value{
Primary basin ID that includes the gauging station. For each line in \code{stations} one primary basin ID will be returned.
}
\description{
Extract primary basin ID for one or multiple gauging stations.
With the primary basin ID the upstream sub-basins are identified.
}
\details{
Stations must be a \code{data.frame} with one or more gauging stations (\code{rows}) with \code{x} (longitude, WGS84) and \code{y} (latitude, WGS84) in two \code{columns}.
}
\examples{
# will follow
}
